requirements.txt
langchain>=0.3.4
pdf2image==1.17.0
boto3==1.34.162
langchain-aws>=0.1.0
opensearch-py~=2.7.1
langchain-community>=0.2.17
requests-aws4auth~=1.3.1
python-dotenv~=1.0.1
langchain-ollama~=0.2.0
ollama~=0.3.3